backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.12
pydantic==2.9.0
pydantic-settings==2.5.0
pymupdf==1.24.11
pdfplumber==0.11.4
pikepdf==9.3.0
cffi==1.17.1
pillow==10.4.0
pytesseract==0.3.13
ocrmypdf==16.5.0
openai==1.47.0
httpx==0.27.2
structlog==24.4.0
aiofiles==24.1.0
reportlab==4.2.5
jinja2==3.1.4
